Updated packages and improved handling
Updated all packages and fixed code due to some breaking changes in Fyne 2.6 that changed the behavior of Refresh(). Also when connecting to ASRs the commands were executing before it loaded in so I extended the wait time for them to load and improved the output handling to be more reliable.